funtions: Added ability to redirect after an action
Rather than a write event just returning a success message
and requiring you to click on to more useful data, now write
events can simply redirect to the useful data, returning it
along with a notification message confirming the action.

NOTE: Not yet supported by Android app, however in it's
current iteration it will just show the redirected page (but not the
notification).